+++ secondhand/src/main/java/com/codesquad/secondhand/domain/jwt/domain/Jwt.java
@@ -1,4 +1,4 @@
-package com.codesquad.secondhand.domain.jwt;
+package com.codesquad.secondhand.domain.jwt.domain;
 
 import lombok.Getter;
