+++ secondhand/src/test/java/com/codesquad/secondhand/BaseControllerTest.java
@@ -10,8 +10,8 @@ import org.springframework.test.web.servlet.MockMvc;
 import com.codesquad.secondhand.amazon.S3Uploader;
 import com.codesquad.secondhand.domain.image.service.ImageQueryService;
 import com.codesquad.secondhand.domain.image.service.ImageService;
-import com.codesquad.secondhand.domain.jwt.Jwt;
-import com.codesquad.secondhand.domain.jwt.JwtProvider;
+import com.codesquad.secondhand.domain.jwt.domain.Jwt;
+import com.codesquad.secondhand.domain.jwt.domain.JwtProvider;
 import com.codesquad.secondhand.domain.member.service.MemberService;
 import com.codesquad.secondhand.domain.product.service.ProductService;
 import com.codesquad.secondhand.domain.reaction.service.ReactionService;
